refactor(api): seperate launch and step structs
